fix scrolling glitch when anchor link is set
On mobile/touch devices the scroll position was
glitching back to the anchor position when an
id was specified. The reason have been some layout
optimizations in this theme that apparently
didn't work well with the JS of the Doxygen treeview.

The scrolling glitch problem is now fixed, but the
available screen area for displaying content got
smaller.