replace inherit section arrow icon with CSS triangle
The "inherited from" sections in a class documentation are hidden by
default. A small arrow icon indicates that the user can interact with
the given text and expand the items that are hidden behind it.

The icon is an image by default, which comes with two downsides:
1. It's not a vector graphic and looks oldschool on hdpi monitors
2. It cannot be matched exactly to the theme colors

This is why it was replaced by a CSS triangle that comes with a nice
open/close animation as a small bonus! 😊