Increase compatibility with Doxygen 1.9.5
The new Doxygen version comes with some small CSS adjustments
that required fixes in this theme.

Nervertheless it is NOT recommended to update to the new Doxygen
version! Doxygen 1.9.5 comes with scrollbar customizations that don't
work well together with this theme and they can not be
reverted/patched with just CSS.

See https://github.com/doxygen/doxygen/issues/9549