feat: make the project installable with a Makefile
This patch adds a small POSIX Makefile so that Doxygen Awesome can be
easily installed, by running `make install`.

The file works on Linux, BSDs, macOS, and possibly even Windows.

This is mostly useful for distributions, so that they can easily package
this project in a uniform way; in fact, I wrote this because I'm
packaging Doxygen Awesome for Debian :)

A default installation would look like this:

    $ tree usr
    usr
    └── local
       └── share
          └── doxygen-awesome-css
             ├── doxygen-awesome-darkmode-toggle.js
             ├── doxygen-awesome-fragment-copy-button.js
             ├── doxygen-awesome-interactive-toc.js
             ├── doxygen-awesome-paragraph-link.js
             ├── doxygen-awesome-sidebar-only-darkmode-toggle.css
             ├── doxygen-awesome-sidebar-only.css
             └── doxygen-awesome.css